LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/html/infrastructure/urls/terminology-0/document-base-url-initiated-grand-parent.https.window-expected.txt


FAIL Check the baseURL of an about:blank document same-origin with its parent assert_equals: expected "https://localhost:9443/html/infrastructure/urls/terminology-0/document-base-url-initiated-grand-parent.https.window.html" but got "https://localhost:9443/common/dispatcher/executor.html?uuid=child_token#interesting-fragment"
FAIL Check the baseURL of an about:blank document cross-origin with its parent promise_test: Unhandled rejection with value: object "SecurityError: Blocked a frame with origin "https://localhost:9443" from accessing a cross-origin frame. Protocols, domains, and ports must match."
FAIL Check the baseURL of an about:blank document cross-site with its parent promise_test: Unhandled rejection with value: object "SecurityError: Blocked a frame with origin "https://localhost:9443" from accessing a cross-origin frame. Protocols, domains, and ports must match."
